+++ CHANGELOG.md
@@ -21,6 +21,10 @@ Changed:
 - replace environment handler with withenv
 - move archive:* modules into build: stage, closes [#2](https://github.com/julian7/goshipdone/issues/2)
 
+Fixed:
+
+- module registration done explicitly
+
 Removed:
 
 - publish:gitlab and publish:github, as publish:artifact already does that